Clientservice: Move yaml parsing out
In preparation for testing, this change moves the yaml parsing code in a
separate file.
