+++ client/clientservice/src/configuration.hpp
@@ -1,0 +1,25 @@
+// Concord
+//
+// Copyright (c) 2021 VMware, Inc. All Rights Reserved.
+//
+// This product is licensed to you under the Apache 2.0 license (the "License"). You may not use this product except in
+// compliance with the Apache 2.0 License.
+//
+// This product may include a number of subcomponents with separate copyright notices and license terms. Your use of
+// these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE
+// file.
+
+#pragma once
+
+#include <boost/program_options.hpp>
+
+#include "client/concordclient/concord_client.hpp"
+#include "Logger.hpp"
+
+namespace concord::client::clientservice {
+
+void configureConcordClient(concord::client::concordclient::ConcordClientConfig& config,
+                            boost::program_options::variables_map& opts,
+                            logging::Logger& l);
+
+}
